@@ -3,34 +3,50 @@
   "version": "1.0.0",
   "description": "",
   "keywords": [],
-  "main": "src/index.js",
   "dependencies": {
-    "@gapo_ui/components": "0.4.52",
     "@reduxjs/toolkit": "1.9.3",
+    "@types/node": "^20.6.0",
+    "@types/react": "^18.2.21",
+    "@types/react-dom": "^18.2.7",
     "antd": "5.2.2",
     "clsx": "1.2.1",
-    "react": "17.0.2",
-    "react-dom": "17.0.2",
-    "react-redux": "8.0.5",
-    "react-scripts": "4.0.0",
-    "redux-injectors": "2.1.0",
-    "redux-saga": "1.2.2",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "react-redux": "^8.1.2",
+    "react-scripts": "^5.0.1",
+    "redux-injectors": "^2.1.0",
+    "redux-saga": "^1.2.3",
+    "typescript": "^5.2.2",
     "uuid": "9.0.0"
   },
   "devDependencies": {
-    "@babel/runtime": "7.13.8",
-    "typescript": "4.1.3"
+    "@types/uuid": "^9.0.3",
+    "@typescript-eslint/eslint-plugin": "^6.7.0",
+    "@typescript-eslint/parser": "^6.7.0",
+    "eslint": "^8.49.0",
+    "eslint-plugin-react": "^7.33.2",
+    "prettier": "^3.0.3",
+    "sass": "^1.66.1"
   },
   "scripts": {
     "start": "react-scripts start",
     "build": "react-scripts build",
     "test": "react-scripts test --env=jsdom",
-    "eject": "react-scripts eject"
+    "eject": "react-scripts eject",
+    "predeploy": "npm run build",
+    "deploy": "gh-pages -d build"
+  },
+  "eslintConfig": {
+    "extends": [
+      "react-app",
+      "react-app/jest"
+    ]
   },
   "browserslist": [
     ">0.2%",
     "not dead",
     "not ie <= 11",
     "not op_mini all"
-  ]
+  ],
+  "homepage": "https://hellodoli.github.io/random-game"
 }
