@@ -1,8 +1,12 @@
-import React, { useState } from "react";
-import { Modal } from "antd";
+import React, { useState, ReactNode } from 'react'
+import { Modal } from 'antd'
 
-const Modals = ({ children }) => {
-  const [open, setOpen] = useState(true);
+interface Props {
+  children?: ReactNode
+}
+
+const Modals = ({ children }: Props) => {
+  const [open, setOpen] = useState(true)
   return (
     <Modal
       title="Modal 1000px width"
@@ -13,7 +17,7 @@ const Modals = ({ children }) => {
     >
       {children}
     </Modal>
-  );
-};
+  )
+}
 
-export default Modals;
+export default Modals
