@@ -1,23 +1,23 @@
-import { configureStore, StoreEnhancer } from "@reduxjs/toolkit";
-import { createInjectorsEnhancer } from "redux-injectors";
-import createSagaMiddleware from "redux-saga";
-import { createReducer } from "./reducers";
+import { configureStore, StoreEnhancer } from '@reduxjs/toolkit'
+import { createInjectorsEnhancer } from 'redux-injectors'
+import createSagaMiddleware from 'redux-saga'
+import { createReducer } from './reducers'
 
-let reduxSagaMonitorOptions = {};
-const sagaMiddleware = createSagaMiddleware(reduxSagaMonitorOptions);
-const { run: runSaga } = sagaMiddleware;
+const reduxSagaMonitorOptions = {}
+const sagaMiddleware = createSagaMiddleware(reduxSagaMonitorOptions)
+const { run: runSaga } = sagaMiddleware
 
 const enhancers = [
   createInjectorsEnhancer({
     createReducer,
-    runSaga
-  })
-] as StoreEnhancer[];
+    runSaga,
+  }),
+] as StoreEnhancer[]
 
 const store = configureStore({
   reducer: createReducer(),
   // middleware: []
-  enhancers
-});
+  enhancers,
+})
 
-export default store;
+export default store
