@@ -1,22 +1,24 @@
-import React, { memo } from "react";
-import { useSelector } from "react-redux";
-import { moneySelector } from "modules/PercentGame/selectors";
+import React, { memo } from 'react'
+import { useSelector } from 'react-redux'
+import { moneySelector } from 'modules/PercentGame/selectors'
 
-import { Flex } from "@gapo_ui/components";
-import CrownCoin from "components/Icons/Game/CrownCoin";
+import CrownCoin from 'components/Icons/Game/CrownCoin'
 
 const Money = () => {
-  const money = useSelector(moneySelector);
+  const money = useSelector(moneySelector)
   return (
-    <Flex
-      UNSAFE_className="consume-item coins"
-      alignItems="center"
-      marginEnd={10}
+    <div
+      className="consume-item coins"
+      style={{
+        display: 'flex',
+        alignItems: 'center',
+        marginRight: 10,
+      }}
     >
       <CrownCoin size={20} classNames="coin-icon" marginRight={4} />
       <span className="text number-coin">{money}</span>
-    </Flex>
-  );
-};
+    </div>
+  )
+}
 
-export default memo(Money);
+export default memo(Money)
