@@ -1,5 +1,4 @@
 import React, { memo } from 'react'
-import { Flex } from '@gapo_ui/components'
 import Notes from './Notes'
 import Money from './Money'
 import Tickets from './Tickets'
@@ -8,11 +7,18 @@ import './style.scss'
 const GameHeader = () => {
   return (
     <div className="game-header">
-      <Flex UNSAFE_className="consume" alignItems="center" flexGrow={1}>
+      <div
+        className="consume"
+        style={{
+          display: 'flex',
+          alignItems: 'center',
+          flexGrow: 1,
+        }}
+      >
         <Notes />
         <Money />
         <Tickets />
-      </Flex>
+      </div>
     </div>
   )
 }
