@@ -1,8 +1,8 @@
-export interface GradientSet {
-  FROM: string;
-  TO: string;
+export interface GradientColorFromTo {
+  FROM: string
+  TO: string
 }
 
 export interface GradientColorSet {
-  [key: string]: GradientSet;
+  [key: string]: GradientColorFromTo
 }
